RegularizeOutputs

+++ CreateSFiles100_n2.R
@@ -1,8 +1,10 @@
 # 
+source("utilities.R")
 source("Semisupervised.R")
 source("ListScenarios.R")
 source("GSSFile.R")
 source("GSFile.R")
+library(tictoc)
 library(ssh)
 #if (!requireNamespace("googledrive", quietly = TRUE)) {
 #  install.packages("googledrive")
@@ -41,45 +43,56 @@ if(!sys_info["nodename"] == "LAPTOP-ADR3M911")
 
 
 
-ini <- 1
-fin <- n2.100.1
+ini <- n2.100.1 
+fin <- n2.100.2
 fin-ini + 1
 
 tic(paste0("Supervised files fitted :"),as.character(fin-ini))
-mclapply(Scenarios2.100[ini:fin], function(x){
+mclapply(Scenarios100[ini:fin], function(x){
   
   command2 <- "ls /home/pgrad1/2201449s/R/CMN/SFiles"
   SFilename <- str_replace(x,"S_","SV_")
-  FilesProcessed <- capture.output(ssh_exec_wait(my_ssh_session,command2))
-  scp_download(my_ssh_session, paste0("/home/pgrad1/2201449s/R/CMN/Scenarios3/",x),pathScenarios )
+  FilesProcessed <- dir(pathSFiles)
+  
+      if(!sys_info["nodename"] == "LAPTOP-ADR3M911")
+      {
+          FilesProcessed <- capture.output(ssh_exec_wait(my_ssh_session,command2))
+          scp_download(my_ssh_session, paste0("/home/pgrad1/2201449s/R/CMN/Scenarios3/",x),pathScenarios )
+        }
+  
   if(is_empty(intersect(FilesProcessed,SFilename))) 
     {
-      Output <- GenerateSFile_SSH(x,pathScenarios)
+      Output <- GenerateSFile(x,pathScenarios)
       saveRDS(Output,paste0(pathSFiles,SFilename))
-      scp_upload(my_ssh_session,paste0(pathSFiles,SFilename),"/home/pgrad1/2201449s/R/CMN/SFiles/")
-      filePathScenario <- paste0(pathScenarios,x)
-      filePathSFile <- paste0(pathSFiles,SFilename)
       
-      # remote the file containing the scenario
-      
-          if (file.exists(filePathScenario)) 
+      if(!sys_info["nodename"] == "LAPTOP-ADR3M911")
       {
-        # Remove the file
-        file.remove(filePathScenario)
-        cat("File deleted successfully.\n")
+        scp_upload(my_ssh_session,paste0(pathSFiles,SFilename),"/home/pgrad1/2201449s/R/CMN/SFiles/")
+        filePathScenario <- paste0(pathScenarios,x)
+        filePathSFile <- paste0(pathSFiles,SFilename)
+        
+        # remote the file containing the scenario
+        
+        if (file.exists(filePathScenario)) 
+        {
+          # Remove the file
+          file.remove(filePathScenario)
+          cat("File deleted successfully.\n")
         } else {
-            cat("File does not exist.\n")
+          cat("File does not exist.\n")
         }
-      
-      # remote the file containing the generate SFile
-      
-      if (file.exists(filePathSFile)) 
-      {
-        # Remove the file
-        file.remove(filePathSFile)
-        cat("File deleted successfully.\n")
-      } else {
-        cat("File does not exist.\n")
+        
+        # remote the file containing the generate SFile
+        
+        if (file.exists(filePathSFile)) 
+        {
+          # Remove the file
+          file.remove(filePathSFile)
+          cat("File deleted successfully.\n")
+        } else {
+          cat("File does not exist.\n")
+        }
+        
       }
       
             
@@ -91,5 +104,5 @@ mclapply(Scenarios2.100[ini:fin], function(x){
   
   
 }, mc.cores = 1)
+
 toc()
-ssh_disconnect()
